Convert LogWindow to a tool that implements IAutoToolformConfig
